+++ internal/rest/models/subscribe/subscriber.go
@@ -1,0 +1,5 @@
+package subscribe
+
+type Subscriber struct {
+	SubscriberID string `json:"subscriber_id"`
+}